[CID-B-05]: editing Sales entity Dao and Service

+++ ProjectWorkspace/CID/CID-Core/src/main/java/com/cid/beans/SalesDto.java
@@ -6,13 +6,15 @@ import lombok.Data;
 
 @Data
 public class SalesDto {
-	private int id;
-	private int userId;
-	private int teamId;
+	private Long id;
+	private Long userId;
+	private Long teamId;
 	private Date date;
 	private int week;
 	private int eurPerHour;
 	private int paxPerHour;
 	private int eurPerPax;
 	private int callsPerHour;	
+	
+	
 }
