+++ ProjectWorkspace/CID/CID-Core/src/main/java/com/cid/dao/helpers/SalesRowMapper.java
@@ -12,15 +12,15 @@ public class SalesRowMapper implements RowMapper<Sales> {
 	@Override
 	public Sales mapRow(ResultSet rs, int rowNum) throws SQLException {
 		Sales sales = new Sales();
-		sales.setId(rs.getInt("id"));
-		sales.setUserId(rs.getInt("user_id"));
-		sales.setTeamId(rs.getInt("team_id"));
+		sales.setId(rs.getLong("id"));
+		sales.setUserId(rs.getLong("user_id"));		
+		sales.setTeamId(rs.getLong("team_id"));
 		sales.setDate(rs.getDate("date"));
-		sales.setWeek(rs.getInt("week"));
+		sales.setWeek(rs.getInt("week"));		
 		sales.setEurPerHour(rs.getInt("eur_per_hour"));
 		sales.setPaxPerHour(rs.getInt("pax_per_hour"));
 		sales.setEurPerPax(rs.getInt("eur_per_pax"));
-		sales.setCallsPerHour(rs.getInt("calls_per_hour"));		
+		sales.setCallsPerHour(rs.getInt("calls_per_hour"));			
 		return sales;
 	}
