+++ ProjectWorkspace/CID/CID-Core/src/main/java/com/cid/dao/helpers/UserRowMapper.java
@@ -13,9 +13,9 @@ public class UserRowMapper implements RowMapper<User>{
 	@Override
 	public User mapRow(ResultSet rs, int rowNum) throws SQLException {
 		User user = new User();
-		user.setDepartmentId(rs.getInt("department_id"));
-		user.setTeamId(rs.getInt("team_id"));	
-		user.setRoleId(rs.getInt("role_id"));
+		user.setDepartmentId(rs.getLong("department_id"));
+		user.setTeamId(rs.getLong("team_id"));	
+		user.setRoleId(rs.getLong("role_id"));
 		user.setFirstname(rs.getString("firstname"));
 		user.setLastname(rs.getString("lastname"));
 		user.setUsername(rs.getString("username"));
